django_backend/vms_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.0,<5.4
psycopg2-binary>=2.9,<2.10
Pillow>=9.0,<10.2
django-cors-headers>=3.10,<4.4
django-filter>=23.2,<25.0 # Added this line
# For production, consider adding gunicorn or uvicorn
# Example: gunicorn>=20.0,<21.0